Dao implementation classes lauren

+++ src/main/java/edu/wpi/cs3733/c22/teamC/controllers/mapEditor/MapNode.java
@@ -1,7 +1,7 @@
 package edu.wpi.cs3733.c22.teamC.controllers.mapEditor;
 
 import edu.wpi.cs3733.c22.teamC.Databases.Location;
-import edu.wpi.cs3733.c22.teamC.Databases.requests.MedicalEquipment;
+import edu.wpi.cs3733.c22.teamC.Databases.MedicalEquipment;
 import edu.wpi.cs3733.c22.teamC.Databases.requests.ServiceRequest;
 import edu.wpi.cs3733.c22.teamC.SQLMethods.LocationQuery;
 import edu.wpi.cs3733.c22.teamC.controllers.ImageLoader;
@@ -40,8 +40,8 @@ public class MapNode extends ImageView {
     this.location =
         new Location(
             LocationQuery.createUniqueNodeID("HALL", MapState.getCurrentFloor()),
-            Double.toString(clickX / MapState.getZoomFactor()),
-            Double.toString(clickY / MapState.getZoomFactor()),
+            Integer.toString((int) (clickX / MapState.getZoomFactor())),
+            Integer.toString((int) (clickY / MapState.getZoomFactor())),
             MapState.getCurrentFloor(),
             "Tower",
             "",
